@@ -22,7 +22,7 @@ module Admin
 
     def update
       @notification = GeneralUserNotification.find(params[:id])
-      if @notification.update_attributes(general_user_notification_parameters)
+      if @notification.update(general_user_notification_parameters)
         flash[:notice] = 'Notification was successfully updated.'
         redirect_to admin_general_user_notifications_path
       else
