@@ -11,20 +11,17 @@ Bundler.require(*Rails.groups)
 
 module Etm
   class Application < Rails::Application
+    # Initialize configuration defaults for originally generated Rails version.
+    config.load_defaults 5.0
+
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration should go into files in config/initializers
-    # -- all .rb files in that directory are automatically loaded.
+    # -- all .rb files in that directory are automatically loaded after loading
+    # the framework and any gems in your application.
 
-    ## Pseudo-modules
-    # I packaged some classes/files separate folders
-    # so we need to load them here. This is only for classes
-    # that belong together but where it didn't make sense to
-    # put them in a module.
-    config.autoload_paths += Dir["#{Rails.root}/lib"]
+    config.autoload_paths += [Rails.root.join('lib').to_s]
 
-    config.active_support.deprecation = :log
     config.time_zone = 'Etc/UTC'
-
     config.encoding = 'utf-8'
 
     config.i18n.available_locales = %i[en nl]
@@ -35,7 +32,7 @@ module Etm
       g.test_framework  :rspec, fixture: false
     end
 
-    # custom 404 and 500 page
+    # Custom 404 and 500 page
     config.exceptions_app = routes
   end
 
