@@ -1,6 +1,6 @@
 # Be sure to restart your server when you modify this file.
 
-# Version of your assets, change this if you want to expire all your assets.
+# Version of your assets, change this if you want to expire all your assets..
 Rails.application.config.assets.version = '1.0'
 
 # Add additional assets to the asset load path
@@ -8,9 +8,8 @@ Rails.application.config.assets.version = '1.0'
 
 # Precompile additional assets.
 # application.js, application.css, and all non-JS/CSS in app/assets folder are already added.
-Rails.application.config.assets.precompile += %w(
+Rails.application.config.assets.precompile += %w[
   admin.css
-  backcasting.js
   compare.js
   establishment_shot.js
   establishment_shot_print.css
@@ -25,7 +24,7 @@ Rails.application.config.assets.precompile += %w(
   report.css
   pico.css
   pico.js
-)
+]
 
 # Compile images in vendor/assets/images
-Rails.application.config.assets.precompile += %w(*.png *.jpg *.jpeg *.gif)
+Rails.application.config.assets.precompile += %w[*.png *.jpg *.jpeg *.gif]
