@@ -2,11 +2,11 @@
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
 #
-# Note that this schema.rb definition is the authoritative source for your
-# database schema. If you need to create the application database on another
-# system, you should be using db:schema:load, not running all the migrations
-# from scratch. The latter is a flawed and unsustainable approach (the more migrations
-# you'll amass, the slower it'll run and the greater likelihood for issues).
+# This file is the source Rails uses to define your schema when running `rails
+# db:schema:load`. When creating a new database, `rails db:schema:load` tends to
+# be faster and is potentially less error prone than running all of your
+# migrations from scratch. Old migrations may fail to apply correctly if those
+# migrations use external dependencies or application code.
 #
 # It's strongly recommended that you check this file into your version control system.
 
@@ -21,13 +21,13 @@ ActiveRecord::Schema.define(version: 2020_10_20_083121) do
   end
 
   create_table "descriptions", id: :integer, options: "ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci", force: :cascade do |t|
-    t.text "content_en", limit: 16777215
+    t.text "content_en", size: :medium
     t.text "short_content_en"
     t.integer "describable_id"
     t.string "describable_type", limit: 191
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.text "content_nl", limit: 16777215
+    t.text "content_nl", size: :medium
     t.text "short_content_nl"
     t.index ["describable_id", "describable_type"], name: "index_descriptions_on_describable_id_and_describable_type"
   end
@@ -90,7 +90,7 @@ ActiveRecord::Schema.define(version: 2020_10_20_083121) do
 
   create_table "sessions", id: :integer, options: "ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci", force: :cascade do |t|
     t.string "session_id", limit: 191, null: false
-    t.text "data", limit: 16777215
+    t.text "data", size: :medium
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.index ["session_id"], name: "index_sessions_on_session_id"
@@ -99,8 +99,8 @@ ActiveRecord::Schema.define(version: 2020_10_20_083121) do
 
   create_table "texts", id: :integer, options: "ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci", force: :cascade do |t|
     t.string "key", limit: 191
-    t.text "content_en", limit: 16777215
-    t.text "content_nl", limit: 16777215
+    t.text "content_en", size: :medium
+    t.text "content_nl", size: :medium
     t.datetime "created_at"
     t.datetime "updated_at"
     t.string "title_en"
